requirements.txt
airsim==1.3.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.19.4
opencv-contrib-python==4.4.0.46
pynput==1.7.1
PySimpleGUI==4.32.1
pyvjoy==1.0.1
six==1.15.0
tornado==4.5.3
